Completed the console view, build model and view switch.
You are now able to switch between the console and editor (Assignment
and Splitscreen not yet implemented). Also created a sort of Builder
model BuilderFactory which takes care of all building of the code,
regardles if this happens on the server or live in Javascript. The
console view is also a bit more pimped, as it now more acurately
reinacts the eclipse console view.