Assignment Switch fix, added business-style of InfoSupport.
Fixed a part of the problem that assignments could not be restarted
after they were started and the user navigated away from the page. The
system now checks for that and gives the possibility to cancel the
assignment or continue with the time theyve already spent given there.
Also added the styling of InfoSupport to the app, not finished, still a
work in progress. Vincent said he had a vision for the assignment
selection pane, we'll see. Styling is definitely not done.